Revert "idk"
This reverts commit f3836075cf28cde04f054343b6cb7dff81f90a7d.

+++ structures.h
@@ -130,8 +130,8 @@ struct Point {
 struct Explosion {
 	float x;
 	float y;
-	float max;
-	float size = 0;
+	int max;
+	int size = 0;
 	std::vector<Point> points = {};
 	Explosion(float x1, float y1, int strength = 33, int n = 32) {
 		x = x1;
